+++ backend/src/main/java/org/example/backend/user/controller/AuthController.java
@@ -1,6 +1,6 @@
-package org.example.backend.controller;
+package org.example.backend.user.controller;
 
-import org.example.backend.model.UserModel;
+import org.example.backend.user.model.UserModel;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.*;
