+++ backend/src/main/java/org/example/backend/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package org.example.backend.repository;
+package org.example.backend.user.repository;
 
-import org.example.backend.model.UserModel;
+import org.example.backend.user.model.UserModel;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
